+++ class/Gini/Controller/CGI/AJAX/Gapper/Client.php
@@ -5,7 +5,7 @@ namespace Gini\Controller\CGI\AJAX\Gapper;
 class Client extends \Gini\Controller\CGI
 {
 
-    use \Gini\Module\RPCTrait;
+    use \Gini\Module\Gapper\Client\RPCTrait;
 
     private function _showJSON($data)
     {
